Revert "Merge branch 'PunchAnalysis' into parameterize_color_chart"
This reverts commit 8ade52bb23791fe083b6d09e1ad958fbb9bd79ae, reversing
changes made to 934d6d69f22352154b81a6b8c750b149f9382c3e.
