+++ README.md
@@ -8,10 +8,10 @@ The script expects that files in the dataset directory follow the pattern PLANTI
 ## Setup and Dependencies
 * Before you start, install: 
     * [Git](https://github.com/git-guides/install-git)
-    * [Anaconda Python](https://www.anaconda.com/download/success)
+    * [Anaconda Python](https://www.anaconda.com/products/individual)
 * Use Git to clone this repository
 * Use Anaconda to install the package scikit-learn
 * Use Anaconda to create and activate a python3 environment for this project
-* Within that environment, use pip to install opencv and rawpy packages
+* Within that environment, use pip to install opencv
     * This project requires OpenCV 4.4+, and the developers have found the most consistent success with installing the pip package opencv-contrib-python
