--- src/color_data/NewChartReflectances.csv
@@ -1,24 +1,0 @@
-90.0,90.0,90.0
-59.1, 59.1, 59.1
-36.2, 36.2, 36.2
-19.8, 19.8, 19.8
-9.0, 9.0, 9.0
-3.1, 3.1, 3.1
-NaN, NaN, Nan
-NaN, NaN, Nan
-NaN, NaN, Nan
-NaN, NaN, Nan
-NaN, NaN, Nan
-NaN, NaN, Nan
-NaN, NaN, Nan
-NaN, NaN, Nan
-NaN, NaN, Nan
-NaN, NaN, Nan
-NaN, NaN, Nan
-NaN, NaN, Nan
-NaN, NaN, Nan
-NaN, NaN, Nan
-NaN, NaN, Nan
-NaN, NaN, Nan
-NaN, NaN, Nan
-NaN, NaN, Nan
